requirements.txt
streamlit==1.28.1
groq>=0.8.0
pydantic==2.5.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0